# backend/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Todo API
  description: |
    REST API for Todo App with gamification features.
    
    ## Features
    - User authentication (JWT-based)
    - Todo CRUD operations with optimistic locking
    - User connections for shared todos
    - Gamification system (XP, levels, badges)
    - Real-time synchronization
    
    ## Performance Targets
    - API Response Time: < 200ms (p95)
    - Authentication: < 100ms (p95)
  version: 1.0.0
  contact:
    name: API Support
    email: api@todoapp.com
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.todoapp.com
    description: Production server
...
paths:
...
  /api/v1/todos:
    get:
      tags:
        - Todos
      summary: List todos
      operationId: listTodos
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, completed]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, urgent]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: List of todos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoListResponse'

    post:
      tags:
        - Todos
      summary: Create a new todo
      operationId: createTodo
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoRequest'
      responses:
        '201':
          description: Todo created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  todo:
                    $ref: '#/components/schemas/Todo'
  /api/v1/todos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags:
        - Todos
      summary: Get a todo by ID
      operationId: getTodo
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Todo retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  todo:
                    $ref: '#/components/schemas/Todo'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Todos
      summary: Update a todo
      operationId: updateTodo
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTodoRequest'
      responses:
        '200':
          description: Todo updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  todo:
                    $ref: '#/components/schemas/Todo'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

    delete:
      tags:
        - Todos
      summary: Delete a todo
      operationId: deleteTodo
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Todo deleted successfully
  /api/v1/todos/{id}/complete:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    post:
      tags:
        - Todos
      summary: Mark a todo as complete
      operationId: completeTodo
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                version:
                  type: integer
              required:
                - version
      responses:
        '200':
          description: Todo marked as complete
          content:
            application/json:
              schema:
                type: object
                properties:
                  todo:
                    $ref: '#/components/schemas/Todo'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Todo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        status:
          type: string
          enum: [pending, in_progress, completed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        created_by:
          type: string
          format: uuid
        assigned_to:
          type: string
          format: uuid
        version:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateTodoRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        priority:
          type: string
          enum: [low, medium, high, urgent]
        assigned_to:
          type: string
          format: uuid
      required:
        - title
    UpdateTodoRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        status:
          type: string
          enum: [pending, in_progress, completed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        assigned_to:
          type: string
          format: uuid
        version:
          type: integer
      required:
        - version
    TodoListResponse:
      type: object
      properties:
        todos:
          type: array
          items:
            $ref: '#/components/schemas/Todo'
        total_count:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'